### 2018-03-15 #618 @lfasnacht
There seem to be an issue with the bullet list in the article proof, does somebody know how can I fix that?
