### 2018-03-15 #618 @arfon
@lfasnacht - this should fix it: https://github.com/UniNE-CHYN/mmappickle/pull/2

The Pandoc converter is a little more fussy than GitHub about markdown structure.
