### 2018-03-19 92c8dfb
Fix pickling of masked arrays with a.mask = False
