### 2018-03-26 #618 @arfon
@stensy, @clemaitre58, @glemaitre - would you be willing to review this submission for JOSS? I'd like to find a couple of reviewers for this submission.
