### 2018-03-27 #651 @whedon
[REVIEW]: mmappickle: A Python 3 library to store memory mappable objects into pickle-compatible files
**Submitting author:** @lfasnacht (<a href="http://orcid.org/0000-0002-9853-8209">Laurent Fasnacht</a>)
**Repository:** <a href="https://github.com/UniNE-CHYN/mmappickle" target ="_blank">https://github.com/UniNE-CHYN/mmappickle</a>
**Version:** v1.0.0
