### 2018-03-27 #618 @arfon
@glemaitre, @clemaitre58 - thanks for agreeing to review this. Please head over to #651 for the actual review.
