### 2018-03-27 UniNE-CHYN/mmappickle#3 @glemaitre
JOSS - Functionality
I will go point by point for the review of the JOSS in the coming day.

But before to start, I got a question asap I saw the project.
What is the difference/benefit of `mmappickle` in comparison with [joblib](https://github.com/joblib/joblib).

joblib is handling the pickiing/unpickling of large numpy array and already manage the parallelization with auto-memmaping for the user. There is also work done on the integration with dask to allow distributed programming. Also joblib is used in scikit-learn and other python library.

So my question is the following: what `mmappickle`offers more than `joblib` regarding the application that you stated in the paper (`multiprocessing` / pickling of numpy array)?

Note:

Regarding the pickling you can refer to: [here](http://gael-varoquaux.info/programming/new_low-overhead_persistence_in_joblib_for_big_data.html) for a blog post regarding the performance at that time (some improvements have been made since this post).

cc @clemaitre58
