### 2018-03-27 #651 @glemaitre
My item linked to the revision:

* [functionality/scope of the project](https://github.com/UniNE-CHYN/mmappickle/issues/3)
