### 2018-03-27 UniNE-CHYN/mmappickle#3 @lfasnacht
Hello!
Thank you very much for agreeing to review. Here's a quick answer, I can get deeper in details if required (but I need to get to know more about `joblib`), let me know.

`joblib` shares multiple goals with `mmappickle`. Here are what I think are the main differences with `joblib`, as far as I know (feel free to correct me if needed):

- `joblib` uses one file per array. This can create issues when working with a lot of arrays simultaneously (limits on the number of open files).
- `joblib` seems to focus mainly on computation. `mmappickle` on the other hand in addition try to ease data storage and distribution. Typically, having all the data in one file makes it simple to share (no need to "zip" files), and in addition the file is using the standard pickle protocol version 4, meaning that the file can be loaded by `pickle.load`. This is in my opinion important when sharing and archiving data, since individual projects, even open source ones, tend to evolve or disappear, while it seems pretty likely that Python pickle will stay quite stable.
- concerning the performance, `mmappickle` has a small, nearly-constant overhead over `numpy.memmap`. Since all the alternatives (including `joblib`) also use `numpy.memmap`, the performance are usually similar. The overhead depend linearly on the number of keys in the dictionary, and is cached, so it is usually not an issue except if the dictionary has a large number (thousands) of keys, and changes frequently.

I hope that make sense!
