### 2018-03-27 UniNE-CHYN/mmappickle#3 @lfasnacht
> Basically joblib.load and joblib.dump are intended for this usage (without multiple files anymore). The compression is coming for free on the fly when dumping:
>
> https://joblib.readthedocs.io/en/latest/persistence.html

Sorry, I was not aware of that, however `mmappickle` is made to store a dictionary of things. If I try to do the same with `joblib`, let say I have the following case:
```python
import joblib, numpy
d={'a':numpy.array([1,2,3]),'b':numpy.array([4,5,6]), 'c': numpy.ma.array([7,8,9],mask=[False, True, False])}
joblib.dump(d, '/tmp/out.pkl')
x=joblib.load('/tmp/out.pkl',mmap_mode='r+')
```

`x['a']` and `x['b']` are memmap, but not `x['c']`. (this would work with `mmappickle`).

Now, how do I add a new key to the dictionary stored in `/tmp/out.pkl`?
