### 2018-03-31 UniNE-CHYN/mmappickle#3 @lfasnacht
> > Now, how do I add a new key to the dictionary stored in /tmp/out.pkl?

> You would need to dump the dict again.

That doesn't seem that efficient, especially as all the matrices have to be written again to disk?

> Could you give more information regarding the use case. Apparently, it seems that having a dictionary is useful for your use-case. It allows to dump and memmap on the fly?

Exactly.

Here's the simplest use case I can think of: hyperspectral time lapses. At every time step, an image is captured, but we don't know how many images will be captured in total. An image is 3D numpy array of a few hundred megabytes usually, and its size is known beforehand (it consists in multiple camera frames). There are lot of advantages to hold all the images in the same files, instead of having a lot of different files (for example all the common metadata are stored only once), and it is not possible to hold all images in RAM simultaneously (usually we use a laptop to capture data).

From the `mmappickle` point of view, here's what happens:

1. at first, the file is created, and the metadata is written.
2. at the beginning of each scan, a new key is added to the dict. This doesn't take too much time as only the matrix structure and a "hole" is written in the file.
3. each frame is written to the (now memmap'ed) array (filling the "hole")
4. then go to step 2, as long as not stopped by the user.

The nice thing of using `mmappickle` is that it is possible to view the content of the file while it is written, or even use it from another program (for example to check if everything looks right, or to run an algorithm on the first captured images while still capturing new data). It is the kind of simultaneous access which is not considered by other libraries.

That's why the common "pickling all data at once"-benchmark doesn't really make sense for `mmappickle`. The common use case it random access (add/delete/memmap) of the dict.

Does that make sense to you?
