### 2018-04-24 87071de
Remove httpdict from master, as it is not production ready
