### 2018-04-24 9620e39
HTTP dictionary initial commit (has to be improved)
