### 2018-04-24 48b9152
Add a __version__ variable, and a test to verify that it complies with PEP 440.
Fixes #5.
