+++ mmappickle/__init__.py
@@ -1,2 +1,3 @@
 from .dict import mmapdict
-__all__ = ['mmapdict']
+from ._version import __version__
+__all__ = ['mmapdict', '__version__']
