### 2018-04-24 4e3b3b0
Rework setup.py, as suggested in #6.
