### 2018-04-24 a51383a
Improved test coverage to 96%.
The remaining 4% are either:

- hard to reach (for example, Python's pickle module doesn't generate PUT for example if PROTO=4)
- not testable directly (__main__ in mmappickle.dict)
- tested, but the coverage plugin doesn't capture it, for example if an importer is written to make `import numpy` throw `ImportError`.

A few improvements in mmappickle.dict were also added, to return meaningful exceptions even when in weird corner cases.

Fixes #11.
