+++ tests/test_dict.py
@@ -31,6 +31,10 @@ class TestDictBase(unittest.TestCase):
             self.assertEqual(m._header.commit_number, 465468)
             self.assertTrue(m._header.is_valid())
             
+            with self.assertRaises(TypeError):
+                m._header.commit_number = 'a'
+            self.assertTrue(m._header.is_valid())
+            
     def test_valid_pickle(self):
         with tempfile.TemporaryFile() as f:
             m = mmapdict(f, picklers = [GenericPickler])
@@ -47,7 +51,104 @@ class TestDictBase(unittest.TestCase):
             del m
                 
         self.assertIsNone(m_ref(), "Reference to object is still valid, something is wrong (using object instance instead of weakref.ref?)")
+
+    def test_nonexistent(self):
+        import os
+        with tempfile.NamedTemporaryFile() as f:
+            os.unlink(f.name)
+            with self.assertRaises(FileNotFoundError):
+                d = mmapdict(f.name, True, picklers = [GenericPickler])
+            #This works and re-create the file
+            d = mmapdict(f.name, False, picklers = [GenericPickler])
+            
+    def test_notafile(self):
+        import os
+        with self.assertRaises(TypeError):
+            d = mmapdict({}, picklers = [GenericPickler])
+            
+            
+    def test_readonly(self):
+        with tempfile.NamedTemporaryFile() as f:
+            m = mmapdict(f, picklers = [GenericPickler])
+            m['test1'] = 234
+            self.assertTrue(m.writable)
+            
+            m2 = mmapdict(f.name, True, picklers = [GenericPickler])
+            self.assertFalse(m2.writable)
+            with self.assertRaises(io.UnsupportedOperation):
+                m2['test1'] = 123            
+            with self.assertRaises(io.UnsupportedOperation):
+                m2['test2'] = 123
+            with self.assertRaises(io.UnsupportedOperation):
+                del m2['test1']
+                
+    def test_convert(self):
+        with tempfile.NamedTemporaryFile() as f:
+            v = {'abc': 123}
+            pickle.dump(v, f)
+            f.flush()
+
+            m = mmapdict(f, picklers = [GenericPickler])
+            assert dict(m) == v
+            
+    def test_put_opcodes(self):
+        import string, itertools
+        d_vals = {}
+        for s in itertools.product(string.ascii_letters, string.ascii_letters):
+            s = ''.join(s)
+            d_vals[s] = s
         
+        with tempfile.TemporaryFile() as f:
+            m = mmapdict(f, picklers = [GenericPickler])
+            for i in range(5):
+                m['test-{}'.format(i)] = d_vals
+            for i in range(5):
+                self.assertDictEqual(m['test-{}'.format(i)], d_vals)
+                
+            
+        
+                
+    def _test_bad_file(self, d, exc=None):
+        with tempfile.NamedTemporaryFile() as f:
+            f.write(d)
+            f.flush()
+            
+            if exc is not None:
+                with self.assertRaises(exc):
+                    m = mmapdict(f.name, picklers = [GenericPickler])
+            else:
+                m = mmapdict(f.name, picklers = [GenericPickler])
+                
+    def test_bad_file_header_1(self):
+        self._test_bad_file(b'\x80', ValueError)
+    def test_bad_file_header_2(self):
+        self._test_bad_file(b'\x81\x04', ValueError)
+    def test_bad_file_header_3(self):
+        self._test_bad_file(b'\x80\x03', ValueError)
+    def test_bad_file_header_4(self):
[…]
