### 2018-04-24 dc5a129
Add test_suite to setup.py
