### 2018-04-24 4252867
Add basic appveyor.yml for Windows CI
