### 2018-04-25 5b4e1ce
Fix unsafe behaviour in test_concurrent_1
