### 2018-04-25 68dfb45
Fix locking on Windows platform
