### 2018-04-25 2874dce
Fix tests in order to have them work on Windows
Windows has different assumption about file access rights on temporary
files.
