### 2018-04-25 a8e1b5f
Python 3.4 is required, and numpy for the tests
