### 2018-04-25 b47fe45
Add pywin32 as a requirement for Windows platforms
