### 2018-04-25 da43d54
Appveyor standard python doesn't provide the required libraries, and cannot install them... trying with miniconda
