### 2018-04-25 3cb839b
Disable AppVeyvor/Miniconda 3.4
On appveyor, numpy with Miniconda 3.4 is broken:

ImportError:
Importing the multiarray numpy extension module failed.  Most
likely you are trying to import a failed build of numpy.
If you're working with a numpy git repo, try `git clean -xdf` (removes all
files not under version control).  Otherwise reinstall numpy.
Original error was: DLL load failed: The specified module could not be found.
