+++ appveyor.yml
@@ -1,17 +1,11 @@
 environment:
   matrix:
-    - PYTHON: "C:\\Miniconda34"
-      PYTHON_VERSION: "3.4"
-      PYTHON_ARCH: "32"
     - PYTHON: "C:\\Miniconda35"
       PYTHON_VERSION: "3.5"
       PYTHON_ARCH: "32"
     - PYTHON: "C:\\Miniconda36"
       PYTHON_VERSION: "3.6"
       PYTHON_ARCH: "32"
-    - PYTHON: "C:\\Miniconda34-x64"
-      PYTHON_VERSION: "3.4"
-      PYTHON_ARCH: "64"
     - PYTHON: "C:\\Miniconda35-x64"
       PYTHON_VERSION: "3.5"
       PYTHON_ARCH: "64"
