### 2018-04-25 eede318
Add AppVeyor badge
