+++ README.md
@@ -20,4 +20,9 @@ Contributing
 
 Please post issues and pull requests on github. Alternatively, you can also send your patches by email.
 
-mmappickle uses travis-ci: [![Build Status](https://travis-ci.org/UniNE-CHYN/mmappickle.svg?branch=master)](https://travis-ci.org/UniNE-CHYN/mmappickle)
+mmappickle tests are run at each commit:
+
+Tool         | Status
+------------ | -------------
+travis-ci | [![Build Status](https://travis-ci.org/UniNE-CHYN/mmappickle.svg?branch=master)](https://travis-ci.org/UniNE-CHYN/mmappickle)
+AppVeyor | [![Build status](https://ci.appveyor.com/api/projects/status/rb3b0s0u6k8vmxfx?svg=true)](https://ci.appveyor.com/project/lfasnacht/mmappickle)
