+++ .travis.yml
@@ -3,7 +3,13 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
+before_install:
+  - pip install pytest pytest-cov
+  - pip install coveralls
 install:
   - pip install .
   - pip install numpy
-script: pytest --ignore tests/test_httpdict.py
+script: 
+  - pytest
+after_success:
+  - coveralls
