### 2018-04-25 dca5130
Add coverage to travis pytest command line
