### 2018-04-25 c146741
Travis: install in editable mode
It seems that coverage doesn't work otherwise...
