+++ .travis.yml
@@ -7,7 +7,7 @@ before_install:
   - pip install pytest pytest-cov
   - pip install coveralls
 install:
-  - pip install .
+  - pip install -e .
   - pip install numpy
 script: 
   - pytest --cov=mmappickle
