### 2018-04-25 eb83ff0
Add coveralls badge
Fixes #13.
