### 2018-04-25 bffdf91
PEP-8 compliance
Mostly by running:

autopep8 --in-place . --max-line-length 160 -r

This fixes #4, and also fixes #6.
