### 2018-04-30 a22c0eb
Documentation improvements
- quickstart has been moved to index
- read version from mmappickle/_version.py
- add a contributing guide
