+++ doc/source/conf.py
@@ -15,6 +15,7 @@
 
 import sys
 import os
+import re
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -59,9 +60,11 @@ copyright = '2018, Laurent Fasnacht'
 # built documents.
 #
 # The short X.Y version.
-version = '1.0.0'
+version_data = open(os.path.join(os.path.dirname(os.path.dirname(os.path.dirname(__file__))), 'mmappickle', '_version.py'), 'r').read()
+mo = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]", version_data, re.M)     
+version = mo.group(1)
 # The full version, including alpha/beta/rc tags.
-release = '1.0.0'
+release = mo.group(1)
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
