+++ doc/source/install.rst
@@ -5,6 +5,16 @@ Installation
 
 ``mmappickle`` can be installed either using ``pip``, or from source.
 
+Since ``mmappickle`` requires ``Python 3``, you should ensure that you're using the correct ``pip`` executable. On some distributions, the executable is named ``pip3``.
+
+It is possible to check the ``Python`` version of ``pip`` using:
+
+.. code-block:: none
+
+    pip -V
+    
+Similarly, you may need to use ``py.test-3`` instead of ``py.test``.
+
 Installing using ``pip``
 ------------------------
 
@@ -12,7 +22,7 @@ To install using ``pip``, simply run:
 
 .. code-block:: none
 
-    pip3 install mmappickle
+    pip install mmappickle
 
 Manual installation from source
 -------------------------------
@@ -23,8 +33,12 @@ To install manually, run the following:
 
     git clone https://github.com/UniNE-CHYN/mmappickle
     cd mmappickle
-    sudo python3 setup.py install
-    
-If root access is not available, simply add ``--user`` to the last command line.
+    pip install .
 
-To contribute, use ``develop`` instead of ``install``. This sets a link to the source folder in the python installation, instead of copying the files.
+To contribute, use ``pip install -e`` instead of ``pip install``. This sets a link to the source folder in the python installation, instead of copying the files.
+
+It is advisable to run the tests to ensure that everything is working. This can be done by running the following command in the ``mmappickle`` directory:
+
+.. code-block:: none
+
+    py.test
