### 2018-04-30 4ae4e0a
Add use cases to the documentation main page
