### 2018-04-30 5c39b60
Improvements to JOSS paper
As suggested by @glemaitre in issue #9.
