+++ doc/joss-paper/paper.bib
@@ -19,6 +19,11 @@
         url = {https://docs.scipy.org/doc/numpy/reference/generated/numpy.memmap.html}
 }
 
+@misc{joblib,
+        title = {{Joblib: running Python functions as pipeline jobs}},
+        url = {https://pythonhosted.org/joblib/}
+}
+
 @misc{pythonpickle,
         title = {{pickle — Python object serialization - Python 3.6.4 documentation}},
         url = {https://docs.python.org/3/library/pickle.html}
