### 2018-05-01 1def988
AppVeyor: test using pytest
