### 2018-05-07 #651 @lfasnacht
@glemaitre - Any news from your side? I think I've addressed all the issues you mentioned, but I'm waiting for your feedback/confirmation to close them, and also to make the 1.0.1 release once your review is finished...

I also think you could update the checkboxes in the checklist above, since a lot of these points have been addressed meanwhile.

Thanks a lot!
