### 2018-05-14 #651 @glemaitre
I find the quick start on the landing page better than in the a separate page.
