+++ doc/source/conf.py
@@ -114,7 +114,7 @@ html_theme = 'default'
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-html_theme_options = html_theme_options = {
+html_theme_options = {
     'navigation_depth': 2
 }
 
