+++ doc/source/toc.rst
@@ -1,0 +1,13 @@
+Documentation contents
+----------------------
+
+
+.. toctree::
+   :maxdepth: 2
+
+   index
+   install
+   api
+   internals
+   contributing
+
