+++ rtd-requirements.txt
@@ -1,1 +1,1 @@
-sphinx-rtd-theme >= 0.3.1
+
