### 2018-05-14 #651 @lfasnacht
@clemaitre58 It was hard, but now it's done. It is indeed much better, thank you very much for the suggestion.
