### 2018-05-29 #651 @arfon
@glemaitre, @clemaitre58 - can you both confirm that we're good to accept here?
