### 2018-05-29 #651 @glemaitre
good on my side                                                                                                                                                                        Sent from my phone - sorry to be brief and potential misspell.                                                                                                                                                                            From: notifications@github.comSent: 29 May 2018 4:00 amTo: joss-reviews@noreply.github.comReply to: reply@reply.github.comCc: g.lemaitre58@gmail.com; mention@noreply.github.comSubject: Re: [openjournals/joss-reviews] [REVIEW]: mmappickle: A Python 3 library to store memory mappable objects into pickle-compatible files (#651)  @glemaitre, @clemaitre58 - can you both confirm that we're good to accept here?

—You are receiving this because you were mentioned.Reply to this email directly, view it on GitHub, or mute the thread.
