### 2018-06-18 #651 @lfasnacht
Dear all,

Thank you very much for your awesome work. I will definitely recommend submitting to JOSS to my colleagues, and will also consider submitting more of my work.

Kind regards,
Laurent
